Fixes example to work with zmq 3.2.3
The ZMQ_HWM constant was split into separate receive and send high water
marks. I've updated the exposed constants to match those in zmq 3.2.3 -
I've also updated the example to use this.

Finally, these options are expressed as int (not u64), so I've updated
the appropriate getters and setters to use the int type. See the docs at
http://api.zeromq.org/3-2:zmq-setsockopt for details.