Add a quickcheck test for z85 encoding/decoding
This test, at least on this system, exposes the bug found and fixed as
part of PR #94 ("Support for Windows"). Note that this change does not
contain the fix, to allow for easier verification that the fix actually
fixes the failing test.

I put the test into its own file, and moved the existing z85 test there,
as well. If we extend the test suite, it would become unwieldly, and a
split by "API area" makes sense IMO.