Don't require a mutable borrow to create sockets
This API change allows sharing a `Context` between threads, while still
being able to create sockets from them.

This is a thing possible in C (obviously), but was forbidden by the Rust
API, due to `Context::socket` taking a mutable self reference. The zmq
context is thread-safe, as noted in zmq(7):

  A ØMQ context is thread safe and may be shared among as many
  application threads as necessary, without any additional locking
  required on the part of the caller.

This means sharing is not an issue with regards to
thread-safety. However, there is an issue due to livetimes: While
dropping a context with live sockets is a defined action, it leads to
immediate deadlocks when sockets are live in the same thread.

This change thus wraps the underlying C context in an `Arc`, and each
socket created from that context will retain a reference via that `Arc`
to avoid that deadlock from happening. Thanks to @tcosprojects for
suggesting that solution.

Note that this change almost certainly causes compiler warnings in client
code, as `Context` does no longer need to be mutable, but should be
quite non-intrusive otherwise; the new reference-counting behavior
should not be visible, except in corner cases as outlined above.

This commit should fix issues #64 and #92.