Don't use zmq_has() in the build script
Rather, detect the relevant features by trying to set socket options.
The other four options do not have to be detected, since their only
effect is whether certain address strings are accepted in the connect()
and bind() calls.

They *could* be detected in the build script, but that would make lead
to the build making connection attempts if the protocols are supported.

The effect here is that we can build again with earlier 0MQ versions,
I've tested down to 3.2.5.