Make message allocation infallible, and reap some benefits
This fixes #118, and allows for implementation of several `From`
implementations for `Message`, which in turn allows for a generic
formulation of `send()` and `send_multipart()` in terms of
`Into<Message>`.

Probably there is similiar potential for genericity for the `recv`
family of methods, but that can be taken care of separately.