Rename `RawSocket` to `RawFd`
I think `RawFd` is a better name for this type, because on Unix, the
file descriptor returned by `Socket::get_fd()` might not be any kind
of socket-like file descriptor: under Linux, for example, you get an
"eventfd". Make clear in the documentation that this is *not* the
underlying file descriptor.

File descriptors under Unix have varying semantics, so `RawFd` does
not violate expectations, in contrast to a integer type called
`RawSocket`. Under Windows, file descriptor is not native terminology,
I think, so no expectations should be present.