Split off `test!` macro into `common` module
This allows us to re-use `test!` in `connection.rs` formerly known as
`shared-context.rs`. The name change reflects a more general purpose
of that file; some slight refactoring was done in light of that.

In addition to adding a timeout to prevent hanging tests, `test!` now
also initializes `env_logger`, which is a new dev-dependency.