Fix `zmq::has` feature detection
Before this commit, `zmq::has` will always return `None` because the
compiler flag ZMQ_HAS_ZMQ_HAS wasn't set at all when feature `zmq_has`
was enabled, and was incorrectly set (typo "rust-cfg" instead of
"rustc-cfg") when the `zmq_has` feature was not enabled, but a
new-enough version of libzmq was detected.

This commit makes the necessary adjustments to the build script, and
adds a test to check that zmq::has is actually being called (i.e. it
returns `Some(_)`).