Make zmq-sys build libzmq on Windows.
This copies a pattern used by many other -sys crates which use a git submodule and build the appropriate native library.

This could probably be used on other platforms, too, but I don't have them available to test right now, and it's less of a problem on a system where you can install libzmq with a package manager anyway.

This still requires that you have CMake installed so that libzmq can use it to build.