Disable build-time feature probing
The feature-probing was introduced originally in commit
d8cb8ee ("Binding update for ZMQ v4.1") in 2016, and was probably
intended to cope with potentially missing C ABI, allowing for rust-zmq
to work with a range of different libzmq versions, including ones that
expose a subset of the C ABI that was supported by rust-zmq. Tying it
to the `zmq_has` runtime probe was a mistake, I realize now, from the
very outset.

Thus, going forward, after resolving this issue, rust-zmq will expose
only functionality supported by the C ABI of the minimum libzmq
version it requires, and will never conditionally disable parts of its
API for older C ABI versions.

Fixes: #275