Update FFI and eliminate magic numbers
This commit regenerates the FFI definitions emitted by `bindgen` and
eliminates all magic numbers from the `zmq` crate, relegating these to
`zmq-sys`, where they belong.

This is an incompatible change for the `zmq-sys` crate, so its version
number is bumped accordingly.

This commit also removes the source-code `link` attribute in `zmq-sys`
for linking against libzmq, providing it via the build script instead,
but only in the non-vendored, non-metadeps case, because otherwise,
the build machinery provides it for us. This should also fix #196 for
older tool chains.

Regarding the conditionals for OpenBSD (EPROTO) and
non-Windows (ENOTSUP); it seems both of them should be supported on
current platforms (checked on OpenBSD 6.5).

In the process, a bunch of API warts are marked with TODO comments,
and a low-hanging fruit has been addressed:

- The API no longer provides any guarantees that enums can be cast to
  integers and used interchangably with their C counterparts. This
  affects `SocketType`, `Mechanism`, and `Error`.

  The ability to cast to integers and get the C enum values was never
  a documented part of the API, so I don't consider this as a breaking
  change.

  Unfortunately, in the case of `SocketEvent`, this "feature" was used
  even in the crates' own "monitor.rs" example, which now serves as a
  unit test. Additionally, the monitoring API is in needs complete
  re-design to make it type-safe anyway, so it has not been touched.