Travis CI: For Rust 1.32.0, only do a compilation check
Since our `dev-dependencies` now require a newer Rust version, no
longer do a full test for Rust 1.32.0. It seems the `time` crate is
especially aggressive in its MSRV bumping; the `dev-dependencies`
should be investigated to see if they can be trimmed to lower the MSRV
for running the test suite.