fix build on 32bit architectures
Some ffi tests depend on the pointer size. Added a variant for 32bit
architectures.

The conditional build for 32bit architectures is intentionally done on
`not(target_pointer_width = "64")` so the test is not silently skipped
on architectures with other pointer sizes. In case somebody really wants
to use zmq-sys on 16 bit architectures, the test will fail and another
variant for that pointer size would have to be added.