Stop circumventing the `std` definition of `FromUtf8Error`
By already taking the `Vec<u8>` out of this error we leave
users no choice to bubble up the error with typical `?` or
`anyhow::Context::context()`, because `vec<u8>` does not implement
`std::error::Error` (it's not an error after all).

Instead of trying to outsmart the standard library and annoying the
user, return the original `FromUtf8Error` *as it is*, leaving the
caller to deal with it however they please (i.e. bubble it up, or
call `into_bytes()` themselves if they wish to continue parsing it
otherwise).